Add support for passing AllowReassociation=False.
Previously boto did not support passing the AllowReassociation parameter
to False in the ec2:AssociateAddress API call.

This is problematic, because the default behavior in a VPC only account
allows automatic reassociation, so boto users had no way to change this.

To preserve compatibility with previous behavior, change the default for
the allow_reassociation argument to None, which will omit the parameter.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/API_AssociateAddress.html